+++ system/platform-core/src/main/java/org/platformlambda/core/models/AsyncMultiInbox.java
@@ -20,7 +20,6 @@ package org.platformlambda.core.models;
 
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
-import org.platformlambda.core.services.DistributedTrace;
 import org.platformlambda.core.system.Platform;
 import org.platformlambda.core.system.EventEmitter;
 import org.platformlambda.core.util.Utility;
@@ -126,33 +125,9 @@ public class AsyncMultiInbox extends InboxBase {
                     executor.submit(() -> holder.promise.complete(result));
                 }
                 if (correlation.to != null && holder.traceId != null && holder.tracePath != null) {
-                    try {
-                        Map<String, Object> payload = new HashMap<>();
-                        Map<String, Object> metrics = new HashMap<>();
-                        metrics.put("origin", Platform.getInstance().getOrigin());
-                        metrics.put("id", holder.traceId);
-                        metrics.put("service", correlation.to);
-                        metrics.put("from", holder.from);
-                        metrics.put("exec_time", reply.getExecutionTime());
-                        metrics.put("round_trip", reply.getRoundTrip());
-                        metrics.put("start", start);
-                        metrics.put("path", holder.tracePath);
-                        payload.put("trace", metrics);
-                        if (!annotations.isEmpty()) {
-                            payload.put(ANNOTATIONS, annotations);
-                        }
-                        metrics.put("status", reply.getStatus());
-                        if (reply.getStatus() >= 400) {
-                            metrics.put("success", false);
-                            metrics.put("exception", reply.getError());
-                        } else {
-                            metrics.put("success", true);
-                        }
-                        EventEnvelope dt = new EventEnvelope().setTo(DistributedTrace.DISTRIBUTED_TRACING);
-                        EventEmitter.getInstance().send(dt.setBody(payload));
-                    } catch (Exception e) {
-                        log.error("Unable to send to {}", DistributedTrace.DISTRIBUTED_TRACING, e);
-                    }
+                    recordRpcTrace(holder.traceId, holder.tracePath, correlation.to, holder.from, start,
+                            reply.getStatus(), reply.getError(),
+                            reply.getExecutionTime(), reply.getRoundTrip(), annotations);
                 }
             }
         }