+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 1. java.sql.Timestamp data type added to SimpleMapper
 2. simple type matching feature is extended with a new string 'concat' method
+3. default REST endpoints for /api/event and actuator services
 
 ### Removed
 
@@ -23,7 +24,7 @@ N/A
 
 ### Changed
 
-N/A
+Sort REST endpoints for orderingly loading
 
 ---
 ## Version 4.2.17, 2/20/2025